+++ app/src/main/AndroidManifest.xml
@@ -2,12 +2,14 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.jinsu.nh_life">
 
-    <uses-permission android:name="android.permission.INTERNET"></uses-permission>
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
+        android:hardwareAccelerated="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:largeHeap="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
@@ -18,18 +20,15 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".activity.MyCouponListActivity"></activity>
+        <activity android:name=".activity.PointListActivity"></activity>
+        <activity android:name=".activity.HealthActivity"></activity>
 
-        <activity android:name=".activity.MyCouponListActivity">
-        </activity>
+        <service
+            android:name=".service.StepCheckService"
+            android:exported="true" />
 
-        <activity android:name=".activity.PointListActivity">
-        </activity>
-
-
-        <service android:name=".service.StepCheckService"
-            android:exported="true"/>
+        <activity android:name=".activity.SendMoneyActivity"></activity>
     </application>
-
-
 
 </manifest>
