+++ app/src/main/java/com/example/jinsu/nh_life/fragment/MissionFragment.java
@@ -1,0 +1,18 @@
+package com.example.jinsu.nh_life.fragment;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.jinsu.nh_life.R;
+
+public class MissionFragment extends Fragment {
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View view = (View) inflater.inflate(R.layout.fragment_tab_mission, container, false);
+
+        return view;
+    }
+}
