+++ app/src/main/res/layout/fragment_tab_mission.xml
@@ -6,6 +6,180 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/layout_main"
-    android:orientation="vertical">
+    android:orientation="vertical"
+    android:background="@color/gray1">
+    
+  <RelativeLayout
+      android:layout_width="match_parent"
+      android:layout_height="476dp"
+      android:layout_marginLeft="7.3dp"
+      android:layout_marginRight="7.3dp"
+      android:layout_marginTop="10dp"
+      android:background="@drawable/mission_background">
+      
+      <FrameLayout
+          android:layout_width="46.7dp"
+          android:layout_height="63.7dp"
+          android:layout_marginLeft="80dp"
+          android:layout_marginTop="70dp">
+
+          <ImageView
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:background="@drawable/mission_ready"/>
+
+          <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="1000"
+              android:textColor="@color/white"
+              android:textSize="11sp"
+              android:fontFamily="@font/yoon330"
+              android:layout_marginTop="38dp"
+              android:layout_marginLeft="11.5dp"/>
+
+      </FrameLayout>
+
+      <FrameLayout
+          android:layout_width="46.7dp"
+          android:layout_height="63.7dp"
+          android:layout_marginLeft="80dp"
+          android:layout_marginTop="70dp">
+
+          <ImageView
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:background="@drawable/mission_ready"/>
+
+          <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="1000"
+              android:textColor="@color/white"
+              android:textSize="11sp"
+              android:fontFamily="@font/yoon330"
+              android:layout_marginTop="38dp"
+              android:layout_marginLeft="10.7dp"/>
+
+      </FrameLayout>
+
+      <FrameLayout
+          android:layout_width="46.7dp"
+          android:layout_height="63.7dp"
+          android:layout_marginLeft="217dp"
+          android:layout_marginTop="70dp">
+
+          <ImageView
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:background="@drawable/mission_ready"/>
+
+          <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="3000"
+              android:textColor="@color/white"
+              android:textSize="11sp"
+              android:fontFamily="@font/yoon330"
+              android:layout_marginTop="38dp"
+              android:layout_marginLeft="10.7dp"/>
+
+      </FrameLayout>
+
+      <FrameLayout
+          android:layout_width="46.7dp"
+          android:layout_height="63.7dp"
+          android:layout_marginLeft="220dp"
+          android:layout_marginTop="198dp">
+
+          <ImageView
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:background="@drawable/mission_ready"/>
+
+          <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="5000"
+              android:textColor="@color/white"
+              android:textSize="11sp"
+              android:fontFamily="@font/yoon330"
+              android:layout_marginTop="38dp"
+              android:layout_marginLeft="10.7dp"/>
+
+      </FrameLayout>
+
+      <FrameLayout
+          android:layout_width="46.7dp"
+          android:layout_height="63.7dp"
+          android:layout_marginLeft="80dp"
+          android:layout_marginTop="198dp">
+
+          <ImageView
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:background="@drawable/mission_ready"/>
+
+          <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="7000"
+              android:textColor="@color/white"
+              android:textSize="11sp"
+              android:fontFamily="@font/yoon330"
+              android:layout_marginTop="38dp"
+              android:layout_marginLeft="10.7dp"/>
+
+      </FrameLayout>
+
+      <FrameLayout
+          android:layout_width="46.7dp"
+          android:layout_height="63.7dp"
+          android:layout_marginLeft="80dp"
+          android:layout_marginTop="328dp">
+
+          <ImageView
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:background="@drawable/mission_ready"/>
+
+          <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="9000"
+              android:textColor="@color/white"
+              android:textSize="11sp"
+              android:fontFamily="@font/yoon330"
+              android:layout_marginTop="38dp"
+              android:layout_marginLeft="10.7dp"/>
+
+      </FrameLayout>
+
+      <FrameLayout
+          android:layout_width="46.7dp"
+          android:layout_height="63.7dp"
+          android:layout_marginLeft="240dp"
+          android:layout_marginTop="328dp">
+
+          <ImageView
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:background="@drawable/mission_ready"/>
+
+          <TextView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:text="10000"
+              android:textColor="@color/white"
+              android:textSize="11sp"
+              android:fontFamily="@font/yoon330"
+              android:layout_marginTop="38dp"
+              android:layout_marginLeft="9dp"/>
+
+      </FrameLayout>
+
+  </RelativeLayout>
+
+
 
 </LinearLayout>
