+++ app/src/main/res/values/strings.xml
@@ -1,12 +1,12 @@
 <resources>
     <string name="app_name">NH_Life</string>
     <string-array name="age">
-        <item >10대</item>
-        <item >20대</item>
-        <item >30대</item>
-        <item >40대</item>
-        <item >50대</item>
-        <item >60대</item>
+        <item>10대</item>
+        <item>20대</item>
+        <item>30대</item>
+        <item>40대</item>
+        <item>50대</item>
+        <item>60대</item>
     </string-array>
 
 </resources>
